+++ services/121-service/src/programs/program/program.service.ts
@@ -113,10 +113,9 @@ export class ProgramService {
       .leftJoinAndSelect('program.customCriteria', 'customCriterium')
       .where('"countryId" = :countryId', { countryId: query });
 
-    const programs = await qb.getMany();
-    if (process.env.NODE_ENV === 'production') {
-      programs.filter(program => program.published);
-    }
+    let programs = await qb.getMany();
+    programs = programs.filter(program => program.published);
+
     const programsCount = programs.length;
     return { programs, programsCount };
   }
@@ -208,11 +207,6 @@ export class ProgramService {
     const changedProgram = await this.findOne(programId);
     if (newState === ProgramPhase.registrationValidation) {
       await this.publish(programId);
-    } else if (newState === ProgramPhase.inclusion) {
-      if (process.env.NODE_ENV === 'production') {
-        // Only unpublish in production-mode, as this inhibits developing/testing
-        await this.unpublish(programId);
-      }
     }
     return this.buildProgramRO(changedProgram);
   }